requirements.txt
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
MLB-StatsAPI==1.0.1
pylint==2.7.2
readchar==3.0.3
requests==2.25.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
